+++ routes/admin.js
@@ -2,28 +2,21 @@ const path = require('path');
 
 const express = require('express');
 
-const rootDir = require('../util/path');
-
 const router = express.Router();
 
-const products = [];
+
+
+const productsController=require("../controllers/products");
+const { Module } = require('module');
 
 // /admin/add-product => GET
-router.get('/add-product', (req, res, next) => {
-  res.render('add-product', {
-    pageTitle: 'Add Product',
-    path: '/admin/add-product',
-    formsCSS: true,
-    productCSS: true,
-    activeAddProduct: true
-  });
-});
+router.get('/add-product',productsController.getAddProduct);
+//Refrence to the fun getAddProd in controllers
 
 // /admin/add-product => POST
-router.post('/add-product', (req, res, next) => {
-  products.push({ title: req.body.title });
-  res.redirect('/');
-});
+router.post('/add-product', 
+    productsController.postAddProduct
+);
 
-exports.routes = router;
-exports.products = products;
+module.exports = router;
+//exports.products = products;
